ACCESS_MASK changes and refactoring
