+++ src/AdvApi32.Desktop/AdvApi32.RegistryAccessRights.cs
@@ -3,8 +3,8 @@
 
 namespace PInvoke
 {
+    using static ACCESS_MASK.StandardRight;
     using static Kernel32;
-    using static Kernel32.ACCESS_MASK.StandardRight;
 
     /// <content>
     /// Contains the specific <see cref="ACCESS_MASK"/> for registry keys.
