+++ src/Kernel32.Shared/Kernel32.Shared.projitems
@@ -30,7 +30,6 @@
       <Generator>MSBuild:GenerateCodeFromAttributes</Generator>
     </Compile>
     <Compile Include="$(MSBuildThisFileDirectory)Kernel32Extensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Kernel32+ACCESS_MASK.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)NTStatusException.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Win32Exception.cs" />
   </ItemGroup>
